[Hotfix!] Temporarily add reopened workflow trigger
This temporarily adds a reopened issue workflow trigger so that we can
reprocess the two actions submitted by fay-rush and allcre on the
breaking main branch.

Reopening these issues will allow the Github action to retrigger on the
latest version of main, using the latest working workflow code, so that
the README will be updated again to a working state.